Turn `src` into a proper module and move tests to a separate directory.
